Treat Shiny.addCustomMessageHandler like the 'global' operation that it is
That is, it shouldn't be registered (repeatedly) on renderValue -- just once on page load. Also, to register events properly, we need to keep track of a mapping between source ids and dom ids.
Lastly, registering events in this way requires eventDataWithKey() to not rely on scope (it shouldn't need it anyway)