Fix annotation bug in plotly_build.R
Fixed a bug in which, if the plot only has one annotation but the 'font' attribute has multiple elements (such as color and size), the annotation becomes duplicated. This was due to the fact that plotly_build, when counting annotations, also counted the length of the 'font' attribute list, so if there's only one annotation but 'font' is of length 2, it created a list of 2 identical annotations. 

Now, when counting the length of annotation attributes, any attribute of class 'list' (which I believe should only apply to 'font'), is removed and not counted.